Fix auto reconnect logic
Auto reconnect should only be set to true after a successful connection to a peripheral, and remain true unless there is an explicit and expected disconnection.

Previously auto reconnect defaulted to true initially, but that is no longer logical considering we now have support for connection time outs and cancellation of a pending connection. In those cases we do not want to assume auto reconnect after a known and handle-able failure.