Nil out connecting and connected peripheral references when Bluetooth is powered off
When Bluetooth is powered off while connected to a device, the CB disconnection callback will run first. At that point Bluejay can't tell what causes the disconnection, so it is unexpected, and will trigger an auto reconnect and also put Bluejay in a isConnecting state. Afterwards, however, we will get a did update state callback from CB, then we will know that the Bluetooth has been powered off. At that point where we cancel and empty everything in the queue, we also need to sanity check the connecting and connected peripheral references and nil them out. Otherwise, Bluejay will get stuck in a isConnecting but disconnected state, and future calls to connect will never succeed.